+++ src/main/java/com/dailyon/wishcartservice/cart/repository/CartCustomRepositoryImpl.java
@@ -1,8 +1,8 @@
 package com.dailyon.wishcartservice.cart.repository;
 
 import com.dailyon.wishcartservice.cart.dto.request.UpsertCartRequest;
-import com.dailyon.wishcartservice.cart.entity.Cart;
-import static com.dailyon.wishcartservice.cart.entity.Cart.CartItem;
+import com.dailyon.wishcartservice.cart.document.Cart;
+import static com.dailyon.wishcartservice.cart.document.Cart.CartItem;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
