+++ src/main/java/com/dailyon/wishcartservice/cart/repository/CartRepository.java
@@ -1,6 +1,6 @@
 package com.dailyon.wishcartservice.cart.repository;
 
-import com.dailyon.wishcartservice.cart.entity.Cart;
+import com.dailyon.wishcartservice.cart.document.Cart;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
