+++ src/main/java/com/dailyon/wishcartservice/cart/service/CartService.java
@@ -2,8 +2,8 @@ package com.dailyon.wishcartservice.cart.service;
 
 import com.dailyon.wishcartservice.cart.dto.request.DeleteCartListRequest;
 import com.dailyon.wishcartservice.cart.dto.request.UpsertCartRequest;
-import com.dailyon.wishcartservice.cart.entity.Cart;
-import static com.dailyon.wishcartservice.cart.entity.Cart.CartItem;
+import com.dailyon.wishcartservice.cart.document.Cart;
+import static com.dailyon.wishcartservice.cart.document.Cart.CartItem;
 import com.dailyon.wishcartservice.cart.repository.CartRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
